Disable internal scan properly based on TRST and inst
Redesign internal SFF so that there are two alternate paths
and that normal operation is not affected by any signals. This
allows us to just control a single output mux, which can be
gated with the value of TRST. When the TAP is reset, use normal
operation, otherwise use scan values (when inst is appropriate).
We don't care about the TAP when operating normally anyway.